requirements.txt
langchain
langchain-community
unstructured
pdfminer.six
python-docx
docx2txt
reportlab
datasets
tqdm
sentence-transformers
langchain-huggingface
pinecone-client
python-dotenv
pinecone
langchain-pinecone
langchain 
langchain-core
langchain-community